+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [0.14.0] - 2026-07-15 - Unambiguous mirrored list options
+
 ### Changed
 
 - **Breaking:** Mirrored `Group.list` item options now retain their structural